feat: move fetch and filter logic to list component, add styles
