+++ rs-react-app/src/components/CountriesList/CountriesList.module.css
@@ -1,4 +1,4 @@
-.item {
+.header {
   display: grid;
   grid-template-columns: repeat(4, 1fr);
   gap: 5px;
