+++ rs-react-app/src/types/types.ts
@@ -2,11 +2,6 @@ export interface CountryItemProps {
   country: Country;
 }
 
-export interface CountriesListProps {
-  countries: Country[];
-
-}
-
 export interface FilterProps {
   regions: string[];
   onChange: (e: string) => void;
@@ -16,16 +11,10 @@ export interface Country {
   name: Name;
   population: number;
   region: string;
-  flags: Flags;
+  flag: string;
 }
 
 interface Name {
   common: string;
   official: string;
-}
-
-interface Flags {
-  png: string;
-  svg: string;
-  alt: string;
 }
